Fold raw resource module into parent.
