+++ src/api/raw/mod.rs
@@ -1,7 +1,6 @@
 mod identifier;
 mod include;
-mod resource;
-pub mod relationship;
+mod relationship;
 
 pub use api::get::{RawGet, GetResponse};
 pub use api::index::{RawIndex, IndexResponse};
@@ -11,4 +10,19 @@ pub use api::post::{RawPost, RawPostAsync, PostResponse};
 pub use self::identifier::Identifier;
 pub use self::include::Include;
 pub use self::relationship::{Relationship, RelationshipLinkage, FetchRelationships, UpdateRelationships, NoRelationships};
-pub use self::resource::{RawFetch, RawUpdate, ResourceObject};
+
+use api::Resource;
+
+pub trait RawFetch: Resource {
+    type Relationships: for<'a> FetchRelationships<'a>;
+}
+
+pub trait RawUpdate: RawFetch {
+    type Relationships: UpdateRelationships;
+}
+
+pub struct ResourceObject<T: RawFetch> {
+    pub id: <T as Resource>::Id,
+    pub attributes: T,
+    pub relationships: <T as RawFetch>::Relationships,
+}
